Reorganize source into ACP section folders
Group protocol code under content/, session/, tool-calls/, file-system/, and
agent-plan/, keep the Antigravity backend under agy/, and rename the agent
entry to agent.ts. Import paths are updated only; behavior is unchanged.
