+++ src/agy/db/scan.ts
